# Conflicts:
#	seeds-config/src/main/resources/bootstrap.yml

+++ seeds-chain-api/src/main/java/com/seeds/chain/service/GameItemsService.java
@@ -1,5 +1,6 @@
 package com.seeds.chain.service;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public interface GameItemsService {
@@ -8,7 +9,7 @@ public interface GameItemsService {
 
     BigInteger mintNft(String uri);
 
-    BigInteger gasFees(String uri);
+    BigDecimal gasFees(String uri);
 
     boolean updateNftAttribute(BigInteger tokenId, String newUri);
